+++ lucene-AIP/src/java/org/apache/lucene/index/MultiReader.java
@@ -328,7 +328,7 @@ public class MultiReader extends IndexReader implements Cloneable {
    * AIP change code (AVGL). Return the avgSize of all segments
    */
   @Override
-  public synchronized int avgDocSize() throws IOException{
+  public synchronized float avgDocSize() throws IOException{
       ensureOpen();
       int docNumber = 0;
       long totalDocSizes = 0;
@@ -337,7 +337,7 @@ public class MultiReader extends IndexReader implements Cloneable {
 	  totalDocSizes += subReaders[i].docSizes();
       }
       
-      return (int) totalDocSizes / docNumber;
+      return (float) totalDocSizes / docNumber;
   }
   
   /**
